Renamed all files and folders according to Dart Preferences.
