+++ lib/services/search_service.dart
