+++ lib/starred_documents.dart
@@ -5,7 +5,7 @@ import 'package:hive_flutter/hive_flutter.dart';
 import 'dart:io';
 import 'package:open_file/open_file.dart';
 import 'package:share/share.dart';
-import 'MainDrawer.dart';
+import 'main_drawer.dart';
 
 class Starred extends StatefulWidget {
   @override
@@ -20,15 +20,13 @@ class _StarredState extends State<Starred> {
       appBar: AppBar(
         title: Text(
           'Starred Docs',
-          style: TextStyle(
-              fontSize: 24),
+          style: TextStyle(fontSize: 24),
         ),
         actions: <Widget>[
           IconButton(
             icon: Icon(Icons.refresh),
             onPressed: () {
-              setState(() {
-              });
+              setState(() {});
             },
           ),
           IconButton(
@@ -40,19 +38,13 @@ class _StarredState extends State<Starred> {
       body: WatchBoxBuilder(
         box: Hive.box('starred'),
         builder: (context, starredBox) {
-          if (starredBox
-              .getAt(0)
-              .length == 0) {
+          if (starredBox.getAt(0).length == 0) {
             return Center(
-              child: Text(
-                  "No PDFs Starred Yet !! "
-              ),
+              child: Text("No PDFs Starred Yet !! "),
             );
           }
           return ListView.builder(
-            itemCount: starredBox
-                .getAt(0)
-                .length,
+            itemCount: starredBox.getAt(0).length,
             itemBuilder: (context, index) {
               return GestureDetector(
                 onTap: () {
@@ -68,44 +60,36 @@ class _StarredState extends State<Starred> {
                       children: [
                         Column(
                           children: [
-                            Text(
-                                starredBox.getAt(0)[index]
-                                    .split('/')
-                                    .last
-                            ),
+                            Text(starredBox.getAt(0)[index].split('/').last),
                           ],
                         ),
                         Row(
                           children: [
                             IconButton(
-                                icon: Icon(
-                                    Icons.share
-                                ),
+                                icon: Icon(Icons.share),
                                 onPressed: () async {
-
-                                  File file = await File(
-                                      starredBox.getAt(0)[index]
-                                  );
+                                  File file =
+                                      await File(starredBox.getAt(0)[index]);
 
                                   final path = file.path;
 
                                   print(path);
 
-                                  Share.shareFiles(
-                                      ['$path'], text: 'Your PDF!');
-                                }
-                            ),
+                                  Share.shareFiles(['$path'],
+                                      text: 'Your PDF!');
+                                }),
                             IconButton(
-                                icon: Icon(
-                                    Icons.star
-                                ),
+                                icon: Icon(Icons.star),
                                 onPressed: () async {
                                   setState(() {
-                                    Hive.box('starred').getAt(0).removeAt(index);
+                                    Hive.box('starred')
+                                        .getAt(0)
+                                        .removeAt(index);
                                   });
-                                  Scaffold.of(context).showSnackBar(SnackBar(content: Text('Removed from starred documents')));
-                                }
-                            )
+                                  Scaffold.of(context).showSnackBar(SnackBar(
+                                      content: Text(
+                                          'Removed from starred documents')));
+                                })
                           ],
                         )
                       ],
